small changed switching to windows
